requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
notion-client>=2.2.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
telethon>=1.37.0
imapclient>=3.0.0
mail-parser>=3.15.0
python-dotenv>=1.0.0
qrcode[pil]>=8.0